chore(dive-log): remove the orphaned MergeDiveDialog and its stale copy
MergeDiveDialog / showMergeDiveDialog have had no callers in lib/ since
15fe6966fb1 dropped both dive-detail entry points. Only the widget's own
test referenced it, so the analyzer stayed quiet and the suite kept a dead
390-line widget green.

Deleting it also retires diveLog_mergeDialog_explanation, which told the
user the fold "can be reversed with 'Unlink computer'". No such action
exists; the reversal is the SOURCES chip menu's Split into separate dive.
That sentence had been translated into every locale, so this removes the
ten diveLog_mergeDialog_* keys from all 11 ARB files (11 wrong strings,
not one) and regenerates app_localizations*.dart.

The deleted test file was not purely a test of dead code: its last group
called the live runDiveConsolidation helper and was the only coverage of
tapping Undo, undo() throwing, and both consolidationErrorText branches,
which combine_dives_dialog_test.dart documented a dependency on. That
coverage moves to run_dive_consolidation_test.dart, which drives the
helper directly and adds the notOverlapping branch the old file never
reached.

Re-wiring the dialog instead (the other option in #1452) would not serve
#552: apply() -> build() -> classify() rejects a non-overlapping pair, and
build()'s offsets reproduce the clock skew rather than correcting it, so
the dialog would fail with "these dives don't overlap" for exactly the
case that motivates it.

Closes #1452
