+++ lib/l10n/arb/app_ar.arb
@@ -1721,16 +1721,6 @@
   "diveLog_consolidate_snackbar": "تم دمج الغوصة كجهاز كمبيوتر إضافي.",
   "diveLog_consolidate_undoError": "تعذّر التراجع عن الدمج.",
   "diveLog_consolidate_undone": "تم التراجع عن الدمج",
-  "diveLog_mergeDialog_confirmSubtitle": "سيتم دمج الغوصة في الساعة {time} مع هذه الغوصة.",
-  "diveLog_mergeDialog_confirmTitle": "تأكيد الدمج",
-  "diveLog_mergeDialog_empty": "لم يتم العثور على غوصات أخرى في هذا اليوم.",
-  "diveLog_mergeDialog_explanation": "سيتم دمج ملف هذه الغوصة والأسطوانات والضغوط والأحداث والوسوم ورفقاء الغوص والمشاهدات في هذه الغوصة كمصدر كمبيوتر إضافي. يمكن التراجع عن هذا الإجراء باستخدام 'إلغاء ربط الكمبيوتر'.",
-  "diveLog_mergeDialog_loadError": "خطأ في تحميل الغوصات: {error}",
-  "diveLog_mergeDialog_merge": "دمج",
-  "diveLog_mergeDialog_next": "التالي",
-  "diveLog_mergeDialog_subtitle": "اختر غوصة من نفس اليوم لدمجها كجهاز كمبيوتر إضافي.",
-  "diveLog_mergeDialog_title": "دمج مع غوصة أخرى",
-  "diveLog_mergeDialog_whatThisDoes": "ما الذي يفعله هذا",
   "diveLog_computerSheet_description": "اختر ملف أي كمبيوتر تريد التحرير منه.",
   "diveLog_computerSheet_title": "اختيار الملف الأولي",
   "diveLog_cylinderSac_avgDepth": "متوسط: {depth}",
