+++ lib/l10n/arb/app_de.arb
@@ -1721,16 +1721,6 @@
   "diveLog_consolidate_snackbar": "Tauchgang als zusätzlicher Computer zusammengeführt.",
   "diveLog_consolidate_undoError": "Die Zusammenführung konnte nicht rückgängig gemacht werden.",
   "diveLog_consolidate_undone": "Zusammenführung rückgängig gemacht",
-  "diveLog_mergeDialog_confirmSubtitle": "Der Tauchgang um {time} wird mit diesem Tauchgang zusammengeführt.",
-  "diveLog_mergeDialog_confirmTitle": "Zusammenführung bestätigen",
-  "diveLog_mergeDialog_empty": "Keine weiteren Tauchgänge an diesem Tag gefunden.",
-  "diveLog_mergeDialog_explanation": "Das Profil, die Flaschen, Drücke, Ereignisse, Tags, Tauchpartner und Sichtungen dieses Tauchgangs werden als zusätzliche Computerquelle in diesen Tauchgang eingefügt. Diese Aktion kann mit „Computer trennen“ rückgängig gemacht werden.",
-  "diveLog_mergeDialog_loadError": "Fehler beim Laden der Tauchgänge: {error}",
-  "diveLog_mergeDialog_merge": "Zusammenführen",
-  "diveLog_mergeDialog_next": "Weiter",
-  "diveLog_mergeDialog_subtitle": "Wähle einen Tauchgang vom selben Tag aus, um ihn als zusätzlichen Computer zusammenzuführen.",
-  "diveLog_mergeDialog_title": "Mit einem anderen Tauchgang zusammenführen",
-  "diveLog_mergeDialog_whatThisDoes": "Was das bewirkt",
   "diveLog_computerSheet_description": "Wähle, von welchem Computerprofil aus bearbeitet wird.",
   "diveLog_computerSheet_title": "Startprofil wählen",
   "diveLog_cylinderSac_avgDepth": "Durchschn.: {depth}",
