+++ lib/l10n/arb/app_en.arb
@@ -2689,30 +2689,6 @@
   "diveLog_consolidate_snackbar": "Dive merged as an additional computer.",
   "diveLog_consolidate_undoError": "Couldn't undo the merge.",
   "diveLog_consolidate_undone": "Merge undone",
-  "diveLog_mergeDialog_confirmSubtitle": "Merging dive at {time} into this dive.",
-  "@diveLog_mergeDialog_confirmSubtitle": {
-    "placeholders": {
-      "time": {
-        "type": "String"
-      }
-    }
-  },
-  "diveLog_mergeDialog_confirmTitle": "Confirm merge",
-  "diveLog_mergeDialog_empty": "No other dives found on this day.",
-  "diveLog_mergeDialog_explanation": "This dive's profile, tanks, pressures, events, tags, buddies, and sightings will be folded into this dive as an additional computer source. This action can be reversed with 'Unlink computer'.",
-  "diveLog_mergeDialog_loadError": "Error loading dives: {error}",
-  "@diveLog_mergeDialog_loadError": {
-    "placeholders": {
-      "error": {
-        "type": "String"
-      }
-    }
-  },
-  "diveLog_mergeDialog_merge": "Merge",
-  "diveLog_mergeDialog_next": "Next",
-  "diveLog_mergeDialog_subtitle": "Select a dive from the same day to merge as an additional computer.",
-  "diveLog_mergeDialog_title": "Merge with another dive",
-  "diveLog_mergeDialog_whatThisDoes": "What this does",
   "diveLog_computerSheet_description": "Select which computer's profile to edit from.",
   "@diveLog_computerSheet_description": {
     "description": "Description of the computer profile selection sheet"
