+++ lib/l10n/arb/app_es.arb
@@ -1721,16 +1721,6 @@
   "diveLog_consolidate_snackbar": "Inmersión fusionada como ordenador adicional.",
   "diveLog_consolidate_undoError": "No se pudo deshacer la fusión.",
   "diveLog_consolidate_undone": "Fusión deshecha",
-  "diveLog_mergeDialog_confirmSubtitle": "Se fusionará la inmersión de las {time} con esta inmersión.",
-  "diveLog_mergeDialog_confirmTitle": "Confirmar fusión",
-  "diveLog_mergeDialog_empty": "No se encontraron otras inmersiones en este día.",
-  "diveLog_mergeDialog_explanation": "El perfil, los tanques, las presiones, los eventos, las etiquetas, los compañeros y los avistamientos de esta inmersión se incorporarán a esta inmersión como un ordenador adicional. Esta acción se puede revertir con 'Desvincular ordenador'.",
-  "diveLog_mergeDialog_loadError": "Error al cargar las inmersiones: {error}",
-  "diveLog_mergeDialog_merge": "Fusionar",
-  "diveLog_mergeDialog_next": "Siguiente",
-  "diveLog_mergeDialog_subtitle": "Selecciona una inmersión del mismo día para fusionarla como ordenador adicional.",
-  "diveLog_mergeDialog_title": "Fusionar con otra inmersión",
-  "diveLog_mergeDialog_whatThisDoes": "Qué hace esto",
   "diveLog_computerSheet_description": "Selecciona desde qué perfil de ordenador editar.",
   "diveLog_computerSheet_title": "Elegir perfil inicial",
   "diveLog_cylinderSac_avgDepth": "Prom: {depth}",
