+++ lib/l10n/arb/app_fr.arb
@@ -1648,16 +1648,6 @@
   "diveLog_consolidate_snackbar": "Plongée fusionnée en tant qu'ordinateur supplémentaire.",
   "diveLog_consolidate_undoError": "Impossible d'annuler la fusion.",
   "diveLog_consolidate_undone": "Fusion annulée",
-  "diveLog_mergeDialog_confirmSubtitle": "La plongée de {time} sera fusionnée avec cette plongée.",
-  "diveLog_mergeDialog_confirmTitle": "Confirmer la fusion",
-  "diveLog_mergeDialog_empty": "Aucune autre plongée trouvée ce jour-là.",
-  "diveLog_mergeDialog_explanation": "Le profil, les blocs, les pressions, les événements, les tags, les binômes et les observations de cette plongée seront intégrés à cette plongée en tant qu'ordinateur supplémentaire. Cette action peut être annulée avec « Dissocier l'ordinateur ».",
-  "diveLog_mergeDialog_loadError": "Erreur lors du chargement des plongées : {error}",
-  "diveLog_mergeDialog_merge": "Fusionner",
-  "diveLog_mergeDialog_next": "Suivant",
-  "diveLog_mergeDialog_subtitle": "Sélectionnez une plongée du même jour à fusionner en tant qu'ordinateur supplémentaire.",
-  "diveLog_mergeDialog_title": "Fusionner avec une autre plongée",
-  "diveLog_mergeDialog_whatThisDoes": "Ce que cela fait",
   "diveLog_computerSheet_description": "Sélectionnez le profil d'ordinateur à partir duquel modifier.",
   "diveLog_computerSheet_title": "Choisir le profil de départ",
   "diveLog_cylinderSac_avgDepth": "Moy : {depth}",
