+++ lib/l10n/arb/app_he.arb
@@ -1648,16 +1648,6 @@
   "diveLog_consolidate_snackbar": "הצלילה מוזגה כמחשב נוסף.",
   "diveLog_consolidate_undoError": "לא ניתן היה לבטל את המיזוג.",
   "diveLog_consolidate_undone": "המיזוג בוטל",
-  "diveLog_mergeDialog_confirmSubtitle": "הצלילה בשעה {time} תמוזג לתוך צלילה זו.",
-  "diveLog_mergeDialog_confirmTitle": "אישור מיזוג",
-  "diveLog_mergeDialog_empty": "לא נמצאו צלילות נוספות ביום זה.",
-  "diveLog_mergeDialog_explanation": "הפרופיל, הבלונים, הלחצים, האירועים, התגיות, השותפים והתצפיות של צלילה זו ימוזגו לתוך צלילה זו כמקור מחשב נוסף. ניתן לבטל פעולה זו באמצעות 'ביטול קישור מחשב'.",
-  "diveLog_mergeDialog_loadError": "שגיאה בטעינת הצלילות: {error}",
-  "diveLog_mergeDialog_merge": "מיזוג",
-  "diveLog_mergeDialog_next": "הבא",
-  "diveLog_mergeDialog_subtitle": "בחר צלילה מאותו היום למיזוג כמחשב נוסף.",
-  "diveLog_mergeDialog_title": "מיזוג עם צלילה אחרת",
-  "diveLog_mergeDialog_whatThisDoes": "מה זה עושה",
   "diveLog_computerSheet_description": "בחר מאיזה פרופיל מחשב לערוך.",
   "diveLog_computerSheet_title": "בחירת פרופיל התחלתי",
   "diveLog_cylinderSac_avgDepth": "ממוצע: {depth}",
