+++ lib/l10n/arb/app_hu.arb
@@ -1648,16 +1648,6 @@
   "diveLog_consolidate_snackbar": "A merülés összevonva egy további számítógépként.",
   "diveLog_consolidate_undoError": "Nem sikerült visszavonni az összevonást.",
   "diveLog_consolidate_undone": "Összevonás visszavonva",
-  "diveLog_mergeDialog_confirmSubtitle": "A(z) {time} időpontban indult merülés összevonásra kerül ezzel a merüléssel.",
-  "diveLog_mergeDialog_confirmTitle": "Összevonás megerősítése",
-  "diveLog_mergeDialog_empty": "Nem található más merülés ezen a napon.",
-  "diveLog_mergeDialog_explanation": "Ennek a merülésnek a profilja, palackjai, nyomásai, eseményei, cimkéi, búvártársai és megfigyelései bekerülnek ebbe a merülésbe mint további számítógépes forrás. Ez a művelet visszavonható a „Számítógép leválasztása” funkcióval.",
-  "diveLog_mergeDialog_loadError": "Hiba a merülések betöltésekor: {error}",
-  "diveLog_mergeDialog_merge": "Összevonás",
-  "diveLog_mergeDialog_next": "Tovább",
-  "diveLog_mergeDialog_subtitle": "Válassz egy merülést ugyanarról a napról, hogy további számítógépként vond össze.",
-  "diveLog_mergeDialog_title": "Összevonás egy másik merüléssel",
-  "diveLog_mergeDialog_whatThisDoes": "Mit csinál ez",
   "diveLog_computerSheet_description": "Válaszd ki, melyik számítógép profiljából szerkesztesz.",
   "diveLog_computerSheet_title": "Kiindulási profil kiválasztása",
   "diveLog_cylinderSac_avgDepth": "Atlag: {depth}",
