+++ lib/l10n/arb/app_it.arb
@@ -1648,16 +1648,6 @@
   "diveLog_consolidate_snackbar": "Immersione unita come computer aggiuntivo.",
   "diveLog_consolidate_undoError": "Impossibile annullare l'unione.",
   "diveLog_consolidate_undone": "Unione annullata",
-  "diveLog_mergeDialog_confirmSubtitle": "L'immersione delle {time} verrà unita a questa immersione.",
-  "diveLog_mergeDialog_confirmTitle": "Conferma unione",
-  "diveLog_mergeDialog_empty": "Nessun'altra immersione trovata in questo giorno.",
-  "diveLog_mergeDialog_explanation": "Il profilo, le bombole, le pressioni, gli eventi, i tag, i compagni e gli avvistamenti di questa immersione verranno incorporati in questa immersione come fonte di un computer aggiuntivo. Questa azione può essere annullata con 'Scollega computer'.",
-  "diveLog_mergeDialog_loadError": "Errore durante il caricamento delle immersioni: {error}",
-  "diveLog_mergeDialog_merge": "Unisci",
-  "diveLog_mergeDialog_next": "Avanti",
-  "diveLog_mergeDialog_subtitle": "Seleziona un'immersione dello stesso giorno da unire come computer aggiuntivo.",
-  "diveLog_mergeDialog_title": "Unisci a un'altra immersione",
-  "diveLog_mergeDialog_whatThisDoes": "Cosa fa questa azione",
   "diveLog_computerSheet_description": "Seleziona da quale profilo del computer modificare.",
   "diveLog_computerSheet_title": "Scegli profilo di partenza",
   "diveLog_cylinderSac_avgDepth": "Media: {depth}",
