+++ lib/l10n/arb/app_nl.arb
@@ -1721,16 +1721,6 @@
   "diveLog_consolidate_snackbar": "Duik samengevoegd als extra computer.",
   "diveLog_consolidate_undoError": "Kon het samenvoegen niet ongedaan maken.",
   "diveLog_consolidate_undone": "Samenvoeging ongedaan gemaakt",
-  "diveLog_mergeDialog_confirmSubtitle": "De duik om {time} wordt samengevoegd met deze duik.",
-  "diveLog_mergeDialog_confirmTitle": "Samenvoegen bevestigen",
-  "diveLog_mergeDialog_empty": "Geen andere duiken gevonden op deze dag.",
-  "diveLog_mergeDialog_explanation": "Het profiel, de flessen, drukken, gebeurtenissen, tags, buddy's en waarnemingen van deze duik worden samengevoegd in deze duik als extra computerbron. Deze actie kan ongedaan worden gemaakt met 'Computer loskoppelen'.",
-  "diveLog_mergeDialog_loadError": "Fout bij het laden van duiken: {error}",
-  "diveLog_mergeDialog_merge": "Samenvoegen",
-  "diveLog_mergeDialog_next": "Volgende",
-  "diveLog_mergeDialog_subtitle": "Selecteer een duik van dezelfde dag om samen te voegen als extra computer.",
-  "diveLog_mergeDialog_title": "Samenvoegen met een andere duik",
-  "diveLog_mergeDialog_whatThisDoes": "Wat dit doet",
   "diveLog_computerSheet_description": "Kies vanaf welk computerprofiel je wilt bewerken.",
   "diveLog_computerSheet_title": "Startprofiel kiezen",
   "diveLog_cylinderSac_avgDepth": "Gem.: {depth}",
