+++ lib/l10n/arb/app_pt.arb
@@ -1721,16 +1721,6 @@
   "diveLog_consolidate_snackbar": "Mergulho mesclado como computador adicional.",
   "diveLog_consolidate_undoError": "Não foi possível desfazer a mesclagem.",
   "diveLog_consolidate_undone": "Mesclagem desfeita",
-  "diveLog_mergeDialog_confirmSubtitle": "O mergulho às {time} será mesclado a este mergulho.",
-  "diveLog_mergeDialog_confirmTitle": "Confirmar mesclagem",
-  "diveLog_mergeDialog_empty": "Nenhum outro mergulho encontrado neste dia.",
-  "diveLog_mergeDialog_explanation": "O perfil, os cilindros, as pressões, os eventos, as tags, os parceiros e os avistamentos deste mergulho serão incorporados a este mergulho como uma fonte de computador adicional. Esta ação pode ser revertida com 'Desvincular computador'.",
-  "diveLog_mergeDialog_loadError": "Erro ao carregar os mergulhos: {error}",
-  "diveLog_mergeDialog_merge": "Mesclar",
-  "diveLog_mergeDialog_next": "Próximo",
-  "diveLog_mergeDialog_subtitle": "Selecione um mergulho do mesmo dia para mesclar como um computador adicional.",
-  "diveLog_mergeDialog_title": "Mesclar com outro mergulho",
-  "diveLog_mergeDialog_whatThisDoes": "O que isso faz",
   "diveLog_computerSheet_description": "Selecione de qual perfil de computador editar.",
   "diveLog_computerSheet_title": "Escolher perfil inicial",
   "diveLog_cylinderSac_avgDepth": "Media: {depth}",
